requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
drf-yasg==1.21.10
python-decouple==3.8
celery==5.5.3
mysqlclient==2.2.7
django-environ==0.12.0
